solitaire: remove recursive FromAccounts and 'c lifetime.
Solitaire parses accounts using two traits, FromAccounts and Peel. The FromAccounts
derive macro generates a function, `FromAccounts::from()`, which calls `Peel::peel`
on each and every field in the struct.

Peel's job here is to verify all the constraints for a field before allowing it to
be used in a program.

FromAccounts also implements `Peel` for the structure itself however, which allows
the struct _itself_ to be used as a field in another struct. This allows programs
to create fields that themselves are made up of multiple Solana Accounts.

This is very confusing to read, and also complicates the codebase as it requires
cleverly managing the list of AccountInfo's solana provides as it's threaded thru
the various layers.

This commit removes the recursion by:

1) Removing the `impl Peel` on accounts derived by FromAccounts
2) Each `Peel::peel()` can now only act on a single account.
3) Removes the AccountInfo iterator from Context as it is no longer needed.
4) Adds the current parsed account to Context instead, safe thanks to (2)
5) Explodes ClaimableVAA into its components everywhere it is used.
6) Removes ClaimableVAA's Peel/FromAccounts impls and uses Peel instead.