solitaire: introduce accounts macro accross all programs
This commit seems large, but it is a mostly mechanical change. It's all in a single
commit in order to make sure the changes still fully compile and run within Tilt.
The changes are as follows:

- Remove `FromAccounts` from solitaire.
- Remove rocksalt entirely, as `FromAccounts` is no longer used.
- Replace `#[derive(FromAccounts)]` everywhere with `accounts!(...)`
- Modifies the `solitaire` macro to call the `accounts!()` functions.

Now that rocksalt and recursion is gone, this change also includes a slight change
to the call to `parse` (formerly `peel`), and passes Context by value. The context
is no longer mutable so a move is more efficient.