cosmwasm: accounting: Don't reference serde::de::Unexpected
The `de::Unexpected` has a `Float(f64)` variant and referencing this
type anywhere causes the compiler to emit the `fmt::Display` impl for
it, which includes an `f64.load` instruction when compiled to wasm,
which causes the contract to be rejected by cosmos chains.

Fix this by using `E::custom` instead of `E::invalid_value` so that we
don't have to reference the `de::Unexpected` type anywhere, preventing
the compiler from emitting its `fmt::Display` impl.